warnings

+++ colorer/src/shared/xml/xmldom.h
@@ -297,7 +297,7 @@ public:
     return null;
   };
 
-  short getNodeType()
+  int getNodeType()
   {
     return type;
   }
